requirements.txt
fastapi
python-multipart~=0.0.9
uvicorn
mongomock
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.32
celery~=5.4.0
numpy~=1.24